backend/requirements.txt
Django==5.0.3
django-clearcache==1.2.1
mysql-connector-python==8.4.0
mysqlclient==2.2.4
python-decouple==3.8
python-dotenv==1.0.0
ldap3==2.9.1